Fixed user deletion (fix #3)

+++ routes/api/jednostka/[jednostkaID]/member/[userID]/remove.js
@@ -14,5 +14,8 @@ export default async function({user, targetJednostka, targetUser}) {
 	const targetFunkcja = await targetUser.getFunkcjaInJednostka(targetJednostka)
 	if(!targetFunkcja) throw new HTTPError(400, "Użytkownik nie jest członkiem jednostki")
 
+	// Ensure user has at least one other funkcja
+	if(!targetUser.isParent && targetUser.funkcje.length <= 1) throw Error("Użytkownik musi mieć przynajmniej jedną funkcję")
+
 	await targetFunkcja.delete()
 }
